application/requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.3
django-appconf==1.0.2
django-compressor==1.6
djangorestframework==3.5.2
drf-nested-routers==0.11.1
static3==0.7.0
